Add barometric sensor feature. Based on I2C. Returns T with better accuracy than the humidity sensor. Update GUI to prettify and add a recording header. An odd problem exists in that the constants obtained in the initialization of the barometric sensor module are not preserved despite being made static.

+++ libs/hardware.h
@@ -41,7 +41,7 @@ void cli(void);
 void sei(void);
 uint32_t check_receive_buffer(void);
 void usart_print_fixed_point(uint32_t value);
-void usart_print_int(int value);
+void usart_print_int(int64_t value);
 void usart_print_hex(uint16_t value);
 void usart_print_string(char *ch);
 void systick_setup(uint16_t period);
